Fix SpendOp ABI: securityZone is uint8, not int32
The Z0tzPrivateLedger contract defines the InEuint64 amount
field with securityZone as uint8 (matching the cofhe-contracts
library). My Vercel relayer port had it as int32, which shifted
every byte offset after securityZone by 3 bytes when encoding
the SpendOp tuple. The contract decoded garbage and reverted
with only ~41K gas consumed.

Local Z0tz/relayer/lib/ledger.ts already had uint8 correctly —
this was a port error in landing/lib/relayer/ledger.ts.

+++ lib/relayer/ledger.ts
@@ -74,7 +74,7 @@ const LEDGER_ABI = [
             type: "tuple",
             components: [
               { name: "ctHash", type: "uint256" },
-              { name: "securityZone", type: "int32" },
+              { name: "securityZone", type: "uint8" },
               { name: "utype", type: "uint8" },
               { name: "signature", type: "bytes" },
             ],
